マイグレーションファイルの修正
誤って作成してしまったtest_camelsテーブルに関する条件分岐を追加

+++ db/migrate/20230613060155_droptest_camels.rb
@@ -1,8 +1,10 @@
 class DroptestCamels < ActiveRecord::Migration[6.0]
   def change
-    drop_table :test_camels do |t|
-      t.references :article, null: false, foreign_key: true
-      t.references :tag, null: false, foreign_key: true
+    if table_exists?(:test_camels)
+      drop_table :test_camels do |t|
+        t.references :article, null: false, foreign_key: true
+        t.references :tag, null: false, foreign_key: true
+      end
     end
   end
 end
